+++ plugins/voxelman/core/events.d
@@ -8,62 +8,40 @@ module voxelman.core.events;
 
 import anchovy.irenderer;
 import dlib.math.vector;
-import tharsis.prof : Profiler;
 
-struct GameStartEvent {
-	Profiler profiler;
-	bool continuePropagation = true;
-}
-struct GameStopEvent {
-	Profiler profiler;
-	bool continuePropagation = true;
-}
+struct GameStartEvent {}
+struct GameStopEvent {}
 
 struct PreUpdateEvent {
 	double deltaTime;
-	Profiler profiler;
-	bool continuePropagation = true;
 }
 struct UpdateEvent {
 	double deltaTime;
-	Profiler profiler;
-	bool continuePropagation = true;
 }
 struct PostUpdateEvent {
 	double deltaTime;
-	Profiler profiler;
-	bool continuePropagation = true;
 }
 
+struct WorldSaveEvent {}
+
 // Initiate drawing in graphics plugin
-struct RenderEvent {
-	Profiler profiler;
-	bool continuePropagation = true;
-}
+struct RenderEvent {}
 
 // draw in 3d. With depth test
 struct Render1Event {
 	IRenderer renderer;
-	Profiler profiler;
-	bool continuePropagation = true;
 }
 
 // draw 2d. without depth test. with alpha
 struct Render2Event {
 	IRenderer renderer;
-	Profiler profiler;
-	bool continuePropagation = true;
 }
 // draw 2d gui. without depth test. with alpha
 struct Render3Event {
 	IRenderer renderer;
-	Profiler profiler;
-	bool continuePropagation = true;
 }
 
 struct WindowResizedEvent
 {
 	uvec2 newSize;
-	Profiler profiler;
-	bool continuePropagation = true;
 }